Introduce a new free-threaded project system API and update shims
This produces a new free-threaded, well factored API for adding projects
to the VisualStudioWorkspace. The core type is the VisualStudioProject
which is a free-threaded API that you can use to push information over.
CSharpProjectShim, VisualBasicProject and CPSProject each have an
isntance of VisualStudioProject that they push things through.

The inheritence model here is now smaller. CSharpProjectShim and
VisualBasicProject inherit from AbstractLegacyProject, but that's it.
CPSProject now inherits nothing, and AbstractProject is here purely
for TypeScript back-compat until they're moved onto the new APIs. The
expectation is F# and TypeScript both move to VisualStudioProject, which
we make public in some form. Then that type will go away.